+++ CHANGELOG.md
@@ -21,6 +21,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   - Add the `replace()` function : Replaces all occurrences of a substring within a string, with optional Unicode (grapheme-safe) and normalization support.
   - Add the `sanitize()` function : Sanitize a string based on configurable flags.
   - Add the `split()` function : Splits a string into an array using a regular expression separator.
+- **Options**
+  - Add the ArrayOption helper
 
 ### Changed
 - **Arrays**